Got even the first day into the standard function.
The catch is that we need to print "\par a partridge..." the very first
time, and "\par and a partridge..." the later times.

The solution is to print "\par \prefix a partridge..." where \prefix is
initially defined to be "" (so the first time it prints as empty), but
immediately after the first time it is used, it gets redefined to "and ".
So effectively it acts as a flag that gets modified after the first time
the prefix used.